+++ src/main/java/org/opensearch/knn/index/engine/KNNLibrary.java
@@ -8,9 +8,11 @@ package org.opensearch.knn.index.engine;
 import org.opensearch.Version;
 import org.opensearch.common.ValidationException;
 import org.opensearch.knn.index.SpaceType;
+import org.opensearch.knn.index.remote.RemoteIndexParameters;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
 /**
  * KNNLibrary is an interface that helps the plugin communicate with k-NN libraries
@@ -149,6 +151,7 @@ public interface KNNLibrary extends MethodResolver {
         return false;
     }
 
+
     /**
      * Checks if the library is deprecated for a given OpenSearch version.
      *
@@ -157,5 +160,10 @@ public interface KNNLibrary extends MethodResolver {
      */
     default boolean isRestricted(Version indexVersionCreated) {
         return false; // By default, libraries are not deprecated
+    }
+
+    default RemoteIndexParameters createRemoteIndexingParameters(Map<String, Object> indexInfoParameters) {
+        throw new UnsupportedOperationException("Remote build service does not support this engine");
+
     }
 }